Add --bind-any-interface command line option to PythonBridge main
This option will bind the msgpack socket server to any interface instead of just to localhost. This means that the server is reachable by any client on the network.
WARNING: this is a security risk and should be used with caution.
The default remains to bind only to localhost and only allow connection from the same machine.

https://github.com/feenkcom/gtoolkit/issues/4172